requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
PyYAML
Pillow
matplotlib
seaborn
pytest
